requirements.txt
fastapi
mangum
uvicorn
transformers
pillow
torch
facenet-pytorch
numpy
opencv-python
